LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/iframe.sub-expected.txt

Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/post-to-owner.py
Blocked access to external URL http://www.127.0.0.1:8800/fetch/metadata/resources/post-to-owner.py


Harness Error (TIMEOUT), message = null

PASS Non-secure same-origin iframe => No headers 
TIMEOUT Non-secure same-site iframe => No headers Test timed out
TIMEOUT Non-secure cross-site iframe => No headers. Test timed out
PASS Secure, cross-site (cross-scheme, same-host) iframe 
PASS Non-secure same-origin iframe => No headers: sec-fetch-dest 
PASS Non-secure same-origin iframe => No headers: sec-fetch-mode 
PASS Non-secure same-origin iframe => No headers: sec-fetch-site 
PASS Non-secure same-origin iframe => No headers: sec-fetch-user 
FAIL Secure, cross-site (cross-scheme, same-host) iframe: sec-fetch-dest assert_equals: expected "iframe" but got ""
FAIL Secure, cross-site (cross-scheme, same-host) iframe: sec-fetch-mode assert_equals: expected "navigate" but got ""
FAIL Secure, cross-site (cross-scheme, same-host) iframe: sec-fetch-site assert_equals: expected "cross-site" but got ""
PASS Secure, cross-site (cross-scheme, same-host) iframe: sec-fetch-user 
